% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{as.data.frame}
\alias{as.data.frame}
\title{a stringsAsFactors=FALSE as.data.frame}
\usage{
as.data.frame(..., stringsAsFactors = FALSE)
}
\arguments{
\item{...}{passed through to data.frame}

\item{stringsAsFactors}{set to false by default}
}
\description{
a stringsAsFactors=FALSE as.data.frame
}
\keyword{internal}

