% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{cleanup_regex_results}
\alias{cleanup_regex_results}
\title{helper for usage of regmatches}
\usage{
cleanup_regex_results(tmp)
}
\arguments{
\item{tmp}{result from regexec or gregexpr or regexpr}
}
\description{
helper for usage of regmatches
}
\keyword{internal}

