% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{cuts_to_spans}
\alias{cuts_to_spans}
\title{helper function that turns cut points into spans}
\usage{
cuts_to_spans(cuts, start = 1, end = Inf)
}
\arguments{
\item{cuts}{where after to cut into pieces}

\item{end}{where does it all end}
}
\description{
helper function that turns cut points into spans
}
\keyword{internal}

