% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{data.frame}
\alias{data.frame}
\title{a stringsAsFactors=FALSE data.frame}
\usage{
data.frame(..., stringsAsFactors = FALSE)
}
\arguments{
\item{...}{passed through to data.frame}

\item{stringsAsFactors}{set to false by default}
}
\description{
a stringsAsFactors=FALSE data.frame
}
\keyword{internal}

