% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{de_sequenize}
\alias{de_sequenize}
\title{helper function to transforms sequences into spans}
\usage{
de_sequenize(x)
}
\arguments{
\item{x}{a bunch of numbers to urn into sequences}
}
\description{
helper function to transforms sequences into spans
}
\keyword{internal}

