% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testing_tools.R
\name{dp_ls}
\alias{dp_ls}
\title{have a look at environments}
\usage{
dp_ls(env = globalenv(), filter = FALSE)
}
\arguments{
\item{env}{environment list objects}

\item{filter}{filter for classes to be returned}
}
\description{
have a look at environments
}

