% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testing_tools.R
\name{dp_tf}
\alias{dp_tf}
\title{text function: wrapper for system.file() to access test files}
\usage{
dp_tf(x = NULL)
}
\arguments{
\item{x}{name of the file}
}
\description{
text function: wrapper for system.file() to access test files
}

