% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{drop_non_group_matches}
\alias{drop_non_group_matches}
\title{helper for usage of regmatches}
\usage{
drop_non_group_matches(tmp, group = TRUE)
}
\arguments{
\item{tmp}{result from regexec or gregexpr or regexpr}
}
\description{
helper for usage of regmatches
}
\keyword{internal}

