% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type.R
\name{fixed}
\alias{fixed}
\alias{regex}
\alias{perl}
\title{Control pattern behaviour}
\usage{
fixed(x)

regex(x, ignore_case = FALSE)

perl(x, ignore_case = FALSE)
}
\arguments{
\item{x}{Pattern whose behaviour will be modified}

\item{ignore_case}{If \code{TRUE}, will ignore case during matching.}
}
\description{
\itemize{
\item \code{perl()}: use \code{perl = TRUE} (the default)
\item \code{regex()}: \code{fixed = FALSE} & \code{perl = FALSE}
\item \code{fixed()}: use \code{fixed = TRUE}
}
}
