% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_match.R
\name{get_groups}
\alias{get_groups}
\title{helper function for text_extract_group}
\usage{
get_groups(x, group)
}
\arguments{
\item{x}{text_extract_group result}

\item{groups}{groups to extract}
}
\description{
helper function for text_extract_group
}
\keyword{internal}

