% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_c.R
\name{\%.\%}
\alias{\%.\%}
\title{concatenating strings operator}
\usage{
a \%.\% b
}
\arguments{
\item{a}{first text}

\item{b}{second text}
}
\description{
concatenating strings operator
}
\seealso{
\link{text_c} (and \link[base]{paste})
}

