% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_concatonate_operator.R
\name{\%..\%}
\alias{\%..\%}
\title{function for concatenating strings}
\usage{
a \%..\% b
}
\arguments{
\item{a}{first string}

\item{b}{first string}
}
\description{
function for concatenating strings
}

