% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{invert_spans}
\alias{invert_spans}
\title{function to invert spans to those numbers not covered}
\usage{
invert_spans(from, to = NULL, start = 1, end = Inf)
}
\arguments{
\item{from}{vector of span starts}

\item{to}{vector of span ends}

\item{start}{minimum}

\item{end}{maximum value}
}
\description{
function to invert spans to those numbers not covered
}

