% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/character_plot.R
\name{plot.character}
\alias{plot.character}
\title{function for plotting text}
\usage{
\method{plot}{character}(x, y = NULL, col = "grey", border = "grey",
  pattern = NULL, pattern_col = "#ED4C4C", ...)
}
\arguments{
\item{x}{object of class rtext}

\item{y}{either NULL or a data.frame with columns "start", "end", "line"}

\item{col}{color for text}

\item{border}{border color for text}

\item{pattern}{regular expression to be searched in text and marked up in plot}

\item{pattern_col}{color for text to be marked up via pattern or y option}

\item{...}{further parameters passed through to text_locate}
}
\description{
function for plotting text
}

