% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{regmatches2}
\alias{regmatches2}
\title{helper function to standardize regexpr results}
\usage{
regmatches2(tmp, group = TRUE)
}
\arguments{
\item{tmp}{regexpr or gregexpr result}
}
\description{
helper function to standardize regexpr results
}
\keyword{internal}

