% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{sequenize}
\alias{sequenize}
\title{helper function to spans into sequences}
\usage{
sequenize(start, end = NULL, simplify = TRUE)
}
\arguments{
\item{start}{first number of sequence}

\item{end}{last number of sequence}

\item{simplify}{discard order, duplicaes etc?}
}
\description{
helper function to spans into sequences
}
\keyword{internal}

