% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{str_extract}
\alias{str_extract}
\alias{str_extract_all}
\title{Extract patterns from a string}
\usage{
str_extract(string, pattern)

str_extract_all(string, pattern)
}
\arguments{
\item{string}{Character vector of strings.}

\item{pattern}{A pattern to match. Can be:
\itemize{
\item A single string or \code{\link[=regex]{regex()}}, to use POSIX 1003.2 extended regular
expression.
\item \code{fixed()} to use a fixed/literal matchi.
\item \code{perl()} to use a Perl-compatible regular expression.
}

Unlike stringr, \code{pattern} must be length 1; it will not be recycled to
the same length as \code{string}.}
}
\description{
Extract patterns from a string
}
