% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match.R
\name{str_match}
\alias{str_match}
\title{Extract matching groups}
\usage{
str_match(string, pattern)
}
\arguments{
\item{string}{Character vector of strings.}

\item{pattern}{A pattern to match. Can be:
\itemize{
\item A single string or \code{\link[=regex]{regex()}}, to use POSIX 1003.2 extended regular
expression.
\item \code{fixed()} to use a fixed/literal matchi.
\item \code{perl()} to use a Perl-compatible regular expression.
}

Unlike stringr, \code{pattern} must be length 1; it will not be recycled to
the same length as \code{string}.}
}
\value{
A character matrix with one row for each element of \code{string}. The first
column gives the complete match; the remaining columns give the component
of each match group (formed by \verb{()}).
}
\description{
Extract matching groups
}
