% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace.R
\name{str_replace}
\alias{str_replace}
\alias{str_replace_all}
\title{Replace matches with a pattern}
\usage{
str_replace(string, pattern, replacement)

str_replace_all(string, pattern, replacement)
}
\arguments{
\item{string}{Character vector of strings.}

\item{pattern}{A pattern to match. Can be:
\itemize{
\item A single string or \code{\link[=regex]{regex()}}, to use POSIX 1003.2 extended regular
expression.
\item \code{fixed()} to use a fixed/literal matchi.
\item \code{perl()} to use a Perl-compatible regular expression.
}

Unlike stringr, \code{pattern} must be length 1; it will not be recycled to
the same length as \code{string}.}

\item{replacement}{A string given replace. References of the form \verb{\\1},
\verb{\\2}, etc will be replaced with the contents of the respective matched
group (created by \verb{()}).}
}
\description{
Replace matches with a pattern
}
