% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sub.R
\name{str_sub}
\alias{str_sub}
\title{Extract substrings}
\usage{
str_sub(string, start = 1L, end = -1L)
}
\arguments{
\item{string}{Character vector of strings.}

\item{start, end}{Integer vectors specifying \code{start} and \code{end} positions
(inclusive). Negative values count backwards from the right hand side;
-1 refers to the last character.

Instead of separate \code{start} and \code{end} parameters you can alternatively
provide a two-column matrix (i.e. from \code{str_locate()}); the first column
will be used for the starting position and the second for the ending
position.}
}
\description{
Extract substrings
}
