% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trim.R
\name{str_trim}
\alias{str_trim}
\title{Trim whitespace off a string}
\usage{
str_trim(string, side = c("both", "left", "right"))
}
\arguments{
\item{string}{Character vector of strings.}

\item{side}{Side on which to remove whitespace (left, right or both).}
}
\description{
Trim whitespace off a string
}
