% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{stringb_arrange}
\alias{stringb_arrange}
\title{function to sort df by variables}
\usage{
stringb_arrange(df, ...)
}
\arguments{
\item{df}{data.frame to be sorted}

\item{...}{column names to use for sorting}
}
\description{
function to sort df by variables
}

