% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{stringb_ls}
\alias{stringb_ls}
\title{have a look at environments}
\usage{
stringb_ls(env = globalenv(), filter = FALSE)
}
\arguments{
\item{env}{environment list objects}

\item{filter}{filter for classes to be returned}
}
\description{
have a look at environments
}

