% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{stringb_tf}
\alias{stringb_tf}
\title{text function: wrapper for system.file() to access test files}
\usage{
stringb_tf(x = NULL)
}
\arguments{
\item{x}{name of the file}
}
\description{
text function: wrapper for system.file() to access test files
}

