% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_c.R
\name{text_c}
\alias{text_c}
\alias{text_c.default}
\title{generic for concatenating strings}
\usage{
text_c(..., sep = "", coll = NULL)

\method{text_c}{default}(..., sep = "", coll = NULL)
}
\arguments{
\item{...}{one or more texts to be concatonated (see also \link[base]{paste})}

\item{sep}{separator between concatonated elements (see also \link[base]{paste})}

\item{coll}{if texts (not only there elements) are to be collapsed as well,
how should the be separated (see also \link[base]{paste})}
}
\description{
generic for concatenating strings

text_c default
}
\seealso{
\link[stringb:grapes-..-grapes]{\%..\%} and \link[stringb:grapes-.-grapes]{\%.\%}
}

