% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_count.R
\name{text_count}
\alias{text_count}
\alias{text_count.default}
\title{generic for counting pattern occurences}
\usage{
text_count(string, pattern, sum = FALSE, vectorize = FALSE, ...)

\method{text_count}{default}(string, pattern, sum = FALSE,
  vectorize = FALSE, ...)
}
\arguments{
\item{string}{text to search through}

\item{pattern}{regex to search for}

\item{sum}{if true all element-wise counts will be summed up}

\item{vectorize}{should function be used in vectorized mode, i.e. should a
pattern with length larger than 1 be allowed and if so, should it be
matched to lines (with recycling if needed) instead of using on element on
all lines}

\item{...}{further arguments passed through to \link[base]{gregexpr}}
}
\description{
generic for counting pattern occurences

text_count defaul method
}

