% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_replace.R
\name{text_delete}
\alias{text_delete}
\alias{text_delete.default}
\title{deleting patterns in string}
\usage{
text_delete(string, pattern = NULL, ...)

\method{text_delete}{default}(string, pattern = NULL, ...)
}
\arguments{
\item{string}{text to be replaced}

\item{pattern}{regex to look for and delete}

\item{...}{further parameter passed through to sub}
}
\description{
deleting patterns in string

deleting patterns in string
}

