% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_detect.R
\name{text_detect}
\alias{text_detect}
\alias{text_detect.default}
\alias{text_grepl}
\title{generic function to test if a regex can be found within a string}
\usage{
text_detect(string, pattern, ...)

\method{text_detect}{default}(string, pattern, ...)

text_grepl(string, pattern, ...)
}
\arguments{
\item{string}{text to be searched through}

\item{pattern}{regex to look for}

\item{...}{further arguments passed through to \link[base]{grepl}}
}
\description{
generic function to test if a regex can be found within a string

text_detect default method

generic function to test if a regex can be found within a string
}

