% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_eval.R
\name{text_eval}
\alias{text_eval}
\title{wrapper function of eval() and parse() to evaluate character vector}
\usage{
text_eval(x, envir = parent.frame(), ...)
}
\arguments{
\item{x}{character vector to be parsed and evaluated}

\item{envir}{where to evaluate character vector}

\item{...}{arguments passed through to eval()}
}
\description{
wrapper function of eval() and parse() to evaluate character vector
}

