% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_extract.R
\name{text_extract}
\alias{text_extract}
\title{extract regex matches}
\usage{
text_extract(x, pattern, ignore.case = FALSE, perl = FALSE, fixed = FALSE,
  useBytes = FALSE, invert = FALSE)
}
\arguments{
\item{x}{text from which to extract}

\item{pattern}{see \link{grep}}

\item{ignore.case}{see \link{grep}}

\item{perl}{see \link{grep}}

\item{fixed}{see \link{grep}}

\item{useBytes}{see \link{grep}}

\item{invert}{if TRUE non-regex-matches are extracted instead}
}
\description{
wrapper function around regexec and regmatches
}

