% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_match (peter-mints in Konflikt stehende Kopie 2016-07-27).R, R/text_match.R
\name{text_extract_group}
\alias{text_extract_group}
\alias{text_extract_group.default}
\title{generic for getting regex group matches}
\usage{
text_extract_group(string, pattern, invert = FALSE, ...)

\method{text_extract_group}{default}(string, pattern, group = NULL,
  invert = FALSE, ...)

text_extract_group(string, pattern, invert = FALSE, ...)

\method{text_extract_group}{default}(string, pattern, group = NULL,
  invert = FALSE, ...)
}
\arguments{
\item{string}{text from which to extract character sequence}

\item{pattern}{regex to be searched for}

\item{...}{further parameter passed through to \link[base]{regexec}}

\item{string}{text from which to extract character sequence}

\item{pattern}{regex to be searched for}

\item{...}{further parameter passed through to \link[base]{regexec}}
}
\description{
generic for getting regex group matches

text default

generic for getting regex group matches

text default
}

