% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_length.R
\name{text_length}
\alias{text_length}
\title{wrapper around nchar to return text length}
\usage{
text_length(x, type = "chars", allowNA = FALSE, keepNA = TRUE,
  na.rm = FALSE)
}
\arguments{
\item{x}{see \link{nchar}}

\item{type}{see \link{nchar}}

\item{allowNA}{see \link{nchar}}

\item{keepNA}{see \link{nchar}}

\item{na.rm}{see \link{nchar}}
}
\description{
wrapper around nchar to return text length
}

