% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_locate.R
\name{text_locate}
\alias{text_locate}
\alias{text_locate.default}
\title{function to get start, end, length form pattern match}
\usage{
text_locate(string, pattern, vectorize = FALSE, ...)

\method{text_locate}{default}(string, pattern, vectorize = FALSE, ...)
}
\arguments{
\item{string}{text to be searched through}

\item{pattern}{regex to look for}

\item{vectorize}{should function be used in vectorized mode, i.e. should a
pattern with length larger than 1 be allowed and if so, should it be
matched to lines (with recycling if needed) instead of using on element on
all lines}

\item{...}{further options passed through to \link[base]{regexpr}}
}
\description{
function to get start, end, length form pattern match

text_locate default
}

