% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_locate.R
\name{text_locate_all}
\alias{text_locate_all}
\alias{text_locate_all.default}
\title{function to get start, end, length form pattern match for all matches}
\usage{
text_locate_all(string, pattern, vectorize = FALSE, simplify = FALSE, ...)

\method{text_locate_all}{default}(string, pattern, vectorize = FALSE,
  simplify = FALSE, ...)
}
\arguments{
\item{string}{text to search through}

\item{pattern}{regex to search for}

\item{vectorize}{should function be used in vectorized mode, i.e. should a
pattern with length larger than 1 be allowed and if so, should it be
matched to lines (with recycling if needed) instead of using on element on
all lines}

\item{simplify}{either getting back a list of results or all list elements
merged into a data.frame with columns identifying original line (i) and
pattern (p) number}

\item{...}{further arguments passed through to \link[base]{gregexpr}}
}
\description{
function to get start, end, length form pattern match for all matches

text_locate_all default
}

