% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_locate.R
\name{text_locate_cleanup}
\alias{text_locate_cleanup}
\title{helper function to standardize regexpr results}
\usage{
text_locate_cleanup(tmp)
}
\arguments{
\item{tmp}{regexpr or gregexpr result}
}
\description{
helper function to standardize regexpr results
}

