% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_locate_group.R
\name{text_locate_group}
\alias{text_locate_group}
\alias{text_locate_group.default}
\title{generic for getting positions regex groups}
\usage{
text_locate_group(string, pattern, group, ...)

\method{text_locate_group}{default}(string, pattern, group, ...)
}
\arguments{
\item{string}{text to be searched through}

\item{pattern}{regex to look for}

\item{group}{integer vector specifying groups to return}

\item{...}{further options passed through to \link[base]{regexpr}}
}
\description{
generic for getting positions regex groups

text_locate_group default
}

