% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_locate.R
\name{text_locate_worker}
\alias{text_locate_worker}
\title{helper function to get start, end, length form pattern match}
\usage{
text_locate_worker(string, pattern, ...)
}
\arguments{
\item{string}{text to be searched through}

\item{pattern}{regex to look for}

\item{...}{further options passed through to \link[base]{regexpr}}
}
\description{
helper function to get start, end, length form pattern match
}

