% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_match (peter-mints in Konflikt stehende Kopie 2016-07-27).R, R/text_match.R
\name{text_match}
\alias{text_match}
\title{generic for getting regex group matches}
\usage{
text_match(string, pattern, invert = FALSE, ...)

text_match(string, pattern, invert = FALSE, ...)
}
\arguments{
\item{string}{text from which to extract character sequence}

\item{pattern}{regex to be searched for}

\item{...}{further parameter passed through to \link[base]{regexpr}}

\item{string}{text from which to extract character sequence}

\item{pattern}{regex to be searched for}

\item{...}{further parameter passed through to \link[base]{regexpr}}
}
\description{
generic for getting regex group matches

generic for getting regex group matches
}
\seealso{
\link{text_extract_group}

\link{text_extract_group}
}

