% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbd_text_replace.R
\name{text_replace_all}
\alias{text_replace_all}
\alias{text_replace_all.default}
\title{replacing patterns in string}
\usage{
text_replace_all(string, pattern = NULL, replacement = NULL, ...)

\method{text_replace_all}{default}(string, pattern = NULL,
  replacement = NULL, ...)
}
\arguments{
\item{string}{text to be replaced}

\item{pattern}{regex to look for}

\item{further}{parameter passed through to gsub}
}
\description{
replacing patterns in string

replacing patterns default
}

