% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_replace_group.R
\name{text_replace_group}
\alias{text_replace_group}
\alias{text_replace_group.default}
\title{function for replacing regex group matches
generic for getting regex group matches}
\usage{
text_replace_group(string, pattern, replacement,
  group = seq_along(replacement), invert = FALSE, ...)

\method{text_replace_group}{default}(string, pattern, replacement,
  group = seq_along(replacement), invert = FALSE, ...)
}
\arguments{
\item{string}{text from which to extract character sequence}

\item{pattern}{regex to be searched for}

\item{...}{further parameter passed through to \link[base]{regexec}}
}
\description{
function for replacing regex group matches
generic for getting regex group matches

text default
}

