% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_replace_group.R
\name{text_replace_locates}
\alias{text_replace_locates}
\alias{text_replace_locates.default}
\title{text_replace_locates default}
\usage{
text_replace_locates(string, found, replacement, group, invert)

\method{text_replace_locates}{default}(string, found, replacement, group,
  invert)
}
\arguments{
\item{string}{text for which to replace parts}

\item{found}{result of an call to text_locate_group or text_locate
- i.e. a list of data.frames
with two columns named 'start' and 'end' that mark character spans
to be replaced within the text elements}

\item{replacement}{character vector of replacements of length 1 or
length(group) to replace regex group matches (marked character spans
provided by the found parameter)}

\item{group}{vector of integers identifying thos regex groups to be replaced}

\item{invert}{should character spans provided by found or their counterparts
be replaced}
}
\description{
text_replace_locates default

text_replace_locates default
}

