% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_snippet.R
\name{text_snippet}
\alias{text_snippet}
\title{retrieving text snippet}
\usage{
text_snippet(x, length = max(nchar(x)), from = NULL, to = NULL,
  coll = FALSE)
}
\arguments{
\item{x}{character vector to be snipped}

\item{length}{length of snippet}

\item{from}{starting character}

\item{to}{last character}

\item{coll}{should a possible vector x with length > 1 collapsed with newline
character as separator?}
}
\description{
function will give back snippets of text via using length,
length and from, length and to, or from and to to specify the snippet
}
\section{Functions}{
\itemize{
\item \code{text_snippet}: retrieving text snippet
}}

