% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_split.R
\name{text_split}
\alias{text_split}
\alias{text_split.default}
\title{generic splitting strings}
\usage{
text_split(string, pattern, vectorize = FALSE, ...)

\method{text_split}{default}(string, pattern, vectorize = FALSE, ...)
}
\arguments{
\item{string}{text to search through}

\item{pattern}{regex to search for}

\item{vectorize}{should function be used in vectorized mode, i.e. should a
pattern with length larger than 1 be allowed and if so, should it be
matched to lines (with recycling if needed) instead of using on element on
all lines}

\item{...}{further arguments passed through to \link[base]{gregexpr}}
}
\description{
generic splitting strings

text_split defaul method
}

