% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_sub.R
\name{text_sub}
\alias{text_sub}
\alias{text_sub.default}
\title{generic for extracting characters sequences by position}
\usage{
text_sub(string, start = NULL, end = NULL)

\method{text_sub}{default}(string, start = NULL, end = NULL)
}
\description{
generic for extracting characters sequences by position

text_sub default
}
\seealso{
\link{text_snippet}
}

