% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_which.R
\name{text_subset}
\alias{text_subset}
\title{generic for subsetting/filtering vectors}
\usage{
text_subset(string, pattern, ...)
}
\arguments{
\item{string}{text to be subsetted}

\item{pattern}{regular expression to subset by}

\item{...}{further arguments passed through to \link[base]{grep}}
}
\description{
generic for subsetting/filtering vectors
}

