% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_to_lower.R
\name{text_to_lower}
\alias{text_to_lower}
\alias{text_to_lower.default}
\title{function for make text lower case}
\usage{
text_to_lower(x)

\method{text_to_lower}{default}(x)
}
\arguments{
\item{x}{text to be processed}
}
\description{
function for make text lower case

default method for text_tolower()
}

