% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_tokenize.R
\name{text_tokenize_lines}
\alias{text_tokenize_lines}
\alias{text_tokenize_lines.default}
\title{generic to tokenize text into lines}
\usage{
text_tokenize_lines(string, non_token = FALSE)

\method{text_tokenize_lines}{default}(string, non_token = FALSE)
}
\arguments{
\item{string}{the text to be tokenized}

\item{non_token}{whether or not token as well as non tokens shall be returned.}
}
\description{
generic to tokenize text into lines

text_tokenize default
}

