% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_tokenize.R
\name{text_tokenize_sentences}
\alias{text_tokenize_sentences}
\alias{text_tokenize_sentences.default}
\title{generic to tokenize text into sentences}
\usage{
text_tokenize_sentences(string, non_token = FALSE)

\method{text_tokenize_sentences}{default}(string, non_token = FALSE)
}
\arguments{
\item{string}{the text to be tokenized}

\item{non_token}{whether or not token as well as non tokens shall be returned.}
}
\description{
generic to tokenize text into sentences

text_tokenize default
}

