% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_trim.R
\name{text_trim}
\alias{text_trim}
\alias{text_trim.default}
\alias{text_trim.list}
\alias{text_trim.numeric}
\title{trim spaces}
\usage{
text_trim(string, side = c("both", "left", "right"), pattern = " ", ...)

\method{text_trim}{default}(string, side = c("both", "left", "right"),
  pattern = " ", ...)

\method{text_trim}{list}(string, side = c("both", "left", "right"),
  pattern = " ", ...)

\method{text_trim}{numeric}(string, side = c("both", "left", "right"),
  pattern = " ", ...)
}
\arguments{
\item{string}{text to be trimmed}

\item{side}{defaults to both might also be left, right, both or b, r, l to
express where to trim pattern away}

\item{pattern}{regex to look for}

\item{...}{further arguments passed through to text_replace()}
}
\description{
trim spaces

trim spaces default

trim spaces list

trim spaces numeric
}

