% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_which.R
\name{text_subset}
\alias{text_filter}
\alias{text_subset}
\title{generic for subsetting/filtering vectors}
\usage{
text_subset(string, pattern, ...)

text_filter(string, pattern, ...)
}
\description{
generic for subsetting/filtering vectors

generic for subsetting/filtering vectors
}

