% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_which.R
\name{text_which}
\alias{text_grep}
\alias{text_which}
\alias{text_which.default}
\title{generic function to know in which elements a pattern can be found}
\usage{
text_which(string, pattern, ...)

\method{text_which}{default}(string, pattern, ...)

text_grep(string, pattern, ...)
}
\arguments{
\item{string}{the text to be searched through}

\item{pattern}{regex to look for}

\item{...}{further arguments passed through to \link[base]{grepl}}
}
\description{
generic function to know in which elements a pattern can be found

text_which default method

generic function to know in which elements a pattern can be found
}

