% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_which.R
\name{text_which_value}
\alias{text_grepv}
\alias{text_which_value}
\alias{text_which_value.default}
\title{generic function to get whole elements in which pattern was found}
\usage{
text_which_value(string, pattern, ...)

text_grepv(string, pattern, ...)

\method{text_which_value}{default}(string, pattern, ...)
}
\arguments{
\item{string}{the character vector to be searched through}

\item{pattern}{regex to look for}

\item{...}{further arguments passed through to \link[base]{grep}}
}
\description{
generic function to get whole elements in which pattern was found

generic function to get whole elements in which pattern was found

text_which_value default method
}

