% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_extract_group.R, R/tools.R
\name{get_groups}
\alias{get_groups}
\title{helper function for text_extract_group}
\usage{
get_groups(x, group)

get_groups(x, group)
}
\arguments{
\item{x}{text_extract_group result}

\item{groups}{groups to extract}

\item{x}{text_replace_group result}

\item{groups}{groups to extract}
}
\description{
helper function for text_extract_group

helper function for text_replace_group
}
\keyword{internal}

