% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_c.R
\name{\%..\%}
\alias{\%..\%}
\title{concatenating strings}
\usage{
a \%..\% b
}
\arguments{
\item{a}{first text}

\item{b}{first text}
}
\description{
concatenating strings
}
\seealso{
\link{text_c} (and \link[base]{paste})
}

