% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect.R
\name{str_detect}
\alias{str_detect}
\title{Detect presence/absence of pattern}
\usage{
str_detect(string, pattern, negate = FALSE)
}
\arguments{
\item{string}{Character vector of strings.}

\item{pattern}{A regular expression.}

\item{negate}{If \code{TRUE}, return non-matching elements.}
}
\description{
Detect presence/absence of pattern
}
