% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_extract_group.R
\name{text_extract_group}
\alias{text_extract_group}
\alias{text_extract_group.default}
\title{generic for getting regex group matches}
\usage{
text_extract_group(string, pattern, group, invert = FALSE, ...)

\method{text_extract_group}{default}(string, pattern, group = NULL,
  invert = FALSE, ...)
}
\arguments{
\item{string}{text from which to extract character sequence}

\item{pattern}{regex to be searched for}

\item{group}{integer vector to indicate those regex group matches to extract}

\item{invert}{whether or no matches or non-matches should be extracted}

\item{...}{further parameter passed through to \link[base]{regexec}}
}
\description{
generic for getting regex group matches

text default
}

