% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_pad.R
\name{text_pad}
\alias{text_pad}
\alias{text_pad.default}
\title{padding text to specified width}
\usage{
text_pad(string, width = max(nchar(string)), pad = " ", side = c("left",
  "right", "both", "l", "r", "b", 1, 2, 3))

\method{text_pad}{default}(string, width = max(nchar(string)), pad = " ",
  side = c("left", "right", "both", "l", "r", "b", 1, 2, 3))
}
\arguments{
\item{string}{text to be wrapped}

\item{width}{width text should have after padding; defaults to: max(nchar(string))}

\item{pad}{the character or character sequence to use for padding}

\item{side}{one of: c("left", "right", "both", "l", "r", "b", 1, 2, 3)}
}
\description{
padding text to specified width

text_wrap default
}

