% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_replace.R
\name{text_replace}
\alias{text_replace}
\alias{text_replace.default}
\title{replacing patterns in string}
\usage{
text_replace(string, pattern = NULL, replacement = NULL, ...)

\method{text_replace}{default}(string, pattern = NULL, replacement = NULL,
  recycle = FALSE, ...)
}
\arguments{
\item{string}{text to be replaced}

\item{pattern}{regex to look for}

\item{replacement}{replacement for pattern found}

\item{...}{further parameter passed through to sub}

\item{recycle}{should arguments be recycled if lengths do not match?}
}
\description{
replacing patterns in string

replacing patterns default
}

