% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_show.R
\name{text_show}
\alias{text_show}
\alias{text_show.default}
\title{showing text}
\usage{
text_show(x, length = 500, from = NULL, to = NULL, coll = FALSE,
  wrap = FALSE, ...)

\method{text_show}{default}(x, length = 500, from = NULL, to = NULL,
  coll = FALSE, wrap = FALSE, ...)
}
\arguments{
\item{x}{text to be shown}

\item{length}{number of characters to be shown}

\item{from}{show from ith character}

\item{to}{show up to ith character}

\item{coll}{should x be collapsed using newline character as binding?}

\item{wrap}{should text be wrapped, or wrapped to certain width, or wrapped
by certain function}

\item{...}{further arguments passed through to \link[base]{cat}}
}
\description{
shows text or portions of the text via cat and the usage of text_snippet()

text_show default
}

