% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_wrap.R
\name{text_wrap}
\alias{text_wrap}
\alias{text_wrap.default}
\title{wraping text to specified width}
\usage{
text_wrap(string, ...)

\method{text_wrap}{default}(string, ...)
}
\arguments{
\item{string}{text to be wrapped}

\item{...}{further arguments passed through to \link[base]{strwrap}}
}
\description{
wraping text to specified width

text_wrap default
}
\seealso{
\link[base]{strwrap}
}

